Fix datatype for long/lat/price and add seeds for orders and reviews

+++ server/database/schema/05_orders.sql
@@ -4,7 +4,7 @@ CREATE TABLE orders (
   customer_id INTEGER REFERENCES users(user_id) ON DELETE CASCADE,
   truck_id INTEGER REFERENCES trucks(truck_id) ON DELETE CASCADE,
   status VARCHAR(255) NOT NULL,
-  total_amount REAL NOT NULL,
+  total_amount DECIMAL(10,2) NOT NULL,
   total_calories INTEGER NOT NULL DEFAULT 0,
   date TIMESTAMP NOT NULL
